Issue 7701: RaisePC public ? NOTE: This can be a breaking change if you were overriding RaisePropertyChanged in your code before.
Issue 7564: The RaisePropertyChanged Overloads Contain Duplicate Code.
This is solved now in ViewModelBase and ObservableObject. All calls are routed
through ObservableObject.RaisePropertyChanged(string). If you override this method only in your VMs,
it will always be called if you call any of the Set methods, RaisePropertyChanged(string)
or RaisePropertyChanged(Expression).